// src/services/types/board-save-request.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const boardSaveRequestTypeEnum = {NOTICE: "NOTICE", BUILD_NOTE: "BUILD_NOTE"} as const;

export type BoardSaveRequestTypeEnumKey = (typeof boardSaveRequestTypeEnum)[keyof typeof boardSaveRequestTypeEnum];

export type BoardSaveRequest = {
	/**
	 * @description 게시판 아이디 - null이면 등록, 아니면 수정
	 * @type integer | undefined, int64
	 */
	id?: number;
	/**
	 * @description 게시판 유형
	 * @type string
	 */
	type: BoardSaveRequestTypeEnumKey;
	/**
	 * @description 등급
	 * @minLength 1
	 * @type string
	 */
	grade: string;
	/**
	 * @description 제목
	 * @minLength 1
	 * @type string
	 */
	title: string;
	/**
	 * @description 내용
	 * @type string | undefined
	 */
	content?: string;
};