// src/services/types/board/board-search-boards.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

import type {ApiResultOnePageBoardPageResponse} from "../api-result-one-page-board-page-response.ts";

export const boardSearchBoardsQueryParamsTypeEnum = {NOTICE: "NOTICE", BUILD_NOTE: "BUILD_NOTE"} as const;

export type BoardSearchBoardsQueryParamsTypeEnumKey =
	(typeof boardSearchBoardsQueryParamsTypeEnum)[keyof typeof boardSearchBoardsQueryParamsTypeEnum];

export type BoardSearchBoardsQueryParams = {
	/**
	 * @description 게시판 유형
	 * @type string | undefined
	 */
	type?: BoardSearchBoardsQueryParamsTypeEnumKey;
	/**
	 * @description 조회 대상 필드 목록
	 * @type array | undefined
	 */
	field?: string[];
	/**
	 * @description 조회 연산자 목록
	 * @type array | undefined
	 */
	operator?: string[];
	/**
	 * @description 조회 값 목록
	 * @type array | undefined
	 */
	value?: string[];
	/**
	 * @type integer | undefined, int32
	 */
	page?: number;
	/**
	 * @type integer | undefined, int32
	 */
	size?: number;
};

/**
 * @description OK
 */
export type BoardSearchBoards200 = ApiResultOnePageBoardPageResponse;

export type BoardSearchBoardsQueryResponse = BoardSearchBoards200;

export type BoardSearchBoardsQuery = {Response: BoardSearchBoards200; QueryParams: BoardSearchBoardsQueryParams; Errors: any};